Time: 7 ms (71.52%), Space: 22.9 MB (44.7%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool helper(vector<int>& piles,int mid,int h){
-        int hours=0;
+        long long hours=0;
         for(int i=0; i<piles.size(); i++){
             hours+=(long)(piles[i]+mid-1)/(long)mid;
         }
